requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    make upgrade
#
asgiref==3.7.2
    # via
    #   -r requirements/quality.txt
    #   django
astroid==3.1.0
    # via
    #   -r requirements/quality.txt
    #   pylint
    #   pylint-celery
build==1.1.1
    # via
    #   -r requirements/pip-tools.txt
    #   pip-tools
cachetools==5.3.3
    # via
    #   -r requirements/ci.txt
    #   tox
chardet==5.2.0
    # via
    #   -r requirements/ci.txt
    #   diff-cover
    #   tox
click==8.1.7
    # via
    #   -r requirements/pip-tools.txt
    #   -r requirements/quality.txt
    #   click-log
    #   code-annotations
    #   edx-lint
    #   pip-tools
click-log==0.4.0
    # via
    #   -r requirements/quality.txt
    #   edx-lint
code-annotations==1.7.0
    # via
    #   -r requirements/quality.txt
    #   edx-lint
colorama==0.4.6
    # via
    #   -r requirements/ci.txt
    #   tox
coverage[toml]==7.4.4
    # via
    #   -r requirements/quality.txt
    #   pytest-cov
diff-cover==8.0.3
    # via -r requirements/dev.in
dill==0.3.8
    # via
    #   -r requirements/quality.txt
    #   pylint
distlib==0.3.8
    # via
    #   -r requirements/ci.txt
    #   virtualenv
django==4.2.11
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/quality.txt
    #   django-crum
    #   django-model-utils
    #   django-mysql
    #   djangorestframework
    #   edx-i18n-tools
    #   jsonfield
    #   openedx-filters
django-crum==0.7.9
    # via -r requirements/quality.txt
django-model-utils==4.4.0
    # via -r requirements/quality.txt
django-mysql==4.12.0
    # via -r requirements/quality.txt
djangorestframework==3.15.0
    # via -r requirements/quality.txt
edx-i18n-tools==1.3.0
    # via -r requirements/dev.in
edx-lint==5.3.6
    # via -r requirements/quality.txt
edx-opaque-keys[django]==2.5.1
    # via -r requirements/quality.txt
eox-tenant==11.0.2
    # via -r requirements/quality.txt
exceptiongroup==1.2.0
    # via
    #   -r requirements/quality.txt
    #   pytest
filelock==3.13.1
    # via
    #   -r requirements/ci.txt
    #   tox
    #   virtualenv
iniconfig==2.0.0
    # via
    #   -r requirements/quality.txt
    #   pytest
isort==5.13.2
    # via
    #   -r requirements/quality.txt
    #   pylint
jinja2==3.1.3
    # via
    #   -r requirements/quality.txt
    #   code-annotations
    #   diff-cover
jsonfield==3.1.0
    # via -r requirements/quality.txt
lxml==5.1.0
    # via edx-i18n-tools
markupsafe==2.1.5
    # via
    #   -r requirements/quality.txt
    #   jinja2
mccabe==0.7.0
    # via
    #   -r requirements/quality.txt
    #   pylint
openedx-filters==1.6.0
    # via -r requirements/quality.txt
packaging==24.0
    # via
    #   -r requirements/ci.txt
    #   -r requirements/pip-tools.txt
    #   -r requirements/quality.txt
    #   build
    #   pyproject-api
    #   pytest
    #   tox
path==16.10.0
    # via edx-i18n-tools
pbr==6.0.0
    # via
    #   -r requirements/quality.txt
    #   stevedore
pip-tools==7.4.1
    # via -r requirements/pip-tools.txt
platformdirs==4.2.0
    # via
    #   -r requirements/ci.txt
    #   -r requirements/quality.txt
    #   pylint
    #   tox
    #   virtualenv
pluggy==1.4.0
    # via
    #   -r requirements/ci.txt
    #   -r requirements/quality.txt
    #   diff-cover
    #   pytest
    #   tox
polib==1.2.0
    # via edx-i18n-tools
pycodestyle==2.11.1
    # via -r requirements/quality.txt
pydocstyle==6.3.0
    # via -r requirements/quality.txt
pygments==2.17.2
    # via diff-cover
pylint==3.1.0
    # via
    #   -r requirements/quality.txt
    #   edx-lint
    #   pylint-celery
    #   pylint-django
    #   pylint-plugin-utils
pylint-celery==0.3
    # via
    #   -r requirements/quality.txt
    #   edx-lint
pylint-django==2.5.5
    # via
    #   -r requirements/quality.txt
    #   edx-lint
pylint-plugin-utils==0.8.2
    # via
    #   -r requirements/quality.txt
    #   pylint-celery
    #   pylint-django
pymongo==3.13.0
    # via
    #   -r requirements/quality.txt
    #   edx-opaque-keys
pyproject-api==1.6.1
    # via
    #   -r requirements/ci.txt
    #   tox
pyproject-hooks==1.0.0
    # via
    #   -r requirements/pip-tools.txt
    #   build
    #   pip-tools
pytest==8.1.1
    # via
    #   -r requirements/quality.txt
    #   pytest-cov
    #   pytest-django
pytest-cov==4.1.0
    # via -r requirements/quality.txt
pytest-django==4.8.0
    # via -r requirements/quality.txt
python-slugify==8.0.4
    # via
    #   -r requirements/quality.txt
    #   code-annotations
pyyaml==6.0.1
    # via
    #   -r requirements/quality.txt
    #   code-annotations
    #   edx-i18n-tools
six==1.16.0
    # via
    #   -r requirements/quality.txt
    #   edx-lint
snowballstemmer==2.2.0
    # via
    #   -r requirements/quality.txt
    #   pydocstyle
sqlparse==0.4.4
    # via
    #   -r requirements/quality.txt
    #   django
stevedore==5.2.0
    # via
    #   -r requirements/quality.txt
    #   code-annotations
    #   edx-opaque-keys
text-unidecode==1.3
    # via
    #   -r requirements/quality.txt
    #   python-slugify
tomli==2.0.1
    # via
    #   -r requirements/ci.txt
    #   -r requirements/pip-tools.txt
    #   -r requirements/quality.txt
    #   build
    #   coverage
    #   pip-tools
    #   pylint
    #   pyproject-api
    #   pyproject-hooks
    #   pytest
    #   tox
tomlkit==0.12.4
    # via
    #   -r requirements/quality.txt
    #   pylint
tox==4.14.1
    # via -r requirements/ci.txt
typing-extensions==4.10.0
    # via
    #   -r requirements/quality.txt
    #   asgiref
    #   astroid
    #   edx-opaque-keys
virtualenv==20.25.1
    # via
    #   -r requirements/ci.txt
    #   tox
wheel==0.43.0
    # via
    #   -r requirements/pip-tools.txt
    #   pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
